requirements.txt
# Azure Databricks Payment Approval Acceleration Demo
# Requirements file

# Core dependencies
databricks-sdk>=0.20.0
pyspark>=3.5.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# ML and modeling
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
mlflow>=2.10.0
hyperopt>=0.2.7

# Feature engineering
feature-engine>=1.6.0
category-encoders>=2.6.0

# Synthetic data generation
faker>=22.0.0
mimesis>=11.1.0

# External data integration
requests>=2.31.0
aiohttp>=3.9.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Delta Lake
delta-spark>=3.0.0

# Streaming
kafka-python>=2.0.2

# API development (for Databricks App)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Monitoring
prometheus-client>=0.19.0

# Date/time handling
python-dateutil>=2.8.2
pytz>=2023.3

# Cryptography (for secure data handling)
cryptography>=42.0.0

# Logging
loguru>=0.7.2
